chore: tidied up some things and fixed mypy errors
